C: Fixed handling of empty CVs
Fixed handling of empty - not successfully autoloaded - CVs (#235, #159, ...)
by storing the SV and XPVCV but ignoring it (and esp. the GV savefileds for it)
and defer to run-time get_cv()

This fixes the !CvCVGV_RC(cv) assertion when autoloading a prev. empty CV
i.e. #235 bytes::length
and also helps the first case in #159
"wrong method dispatch by dumping unfound methods"