C: 1.42_69 add -fdelete-pkg with -O2
control if to ignore compiler dependent packages.
collect deps in the initial Makefile.PL step and store it in B::C::Flags
(not yet tested on windows).
All packages not in @deps are never ignored. This should fix all empty package
problems, such as @ISA, overload or packages without functions