remap improvements via SVp_POK of symbol-name
I used a private patch for Encode 2.48 to store the symbol name in the PVX
slot of the enc IV with SVp_POK to be able to thaw that symbol via dlsym ar run-time
from a frozen image.
This can be uses as perl5 API to store all external XS symbols when they are stored at
compile-time and not properly restored at run-time. I.e. when being initialized compile-time
and used in a lexical or constant.