C: Fix wrong caching of internal LEXWARN symbols #322
fixes t/CORE/op/smartkve.t regression and issue 322.

we cached lexwarn syms twice, and the storage in the global symtable caused
reproducible errors with t/CORE/op/smartkve.t for const->PV and ->IV as
lexwarns, which caused a concat fail on the rhs lexwarn sym.  A lexwarn is
no SV, it is just a special cop_warnings buffer.