C: Support __DATA__ blocks in packages without printing wrong warnings (#310)
Fixes issue 310
Pass a flag to IO::save if it is a __DATA__ fh. So far we only checked for main::DATA.

This fixes the wrong Warning: unable to close filehandle DATA properly when opening
the PerlIO::scalar __DATA__ fh a second time, and also the wrong
Read BEGIN-block $fullname from FileHandle $iotype warning.