C 1.45_13: initialize init0 for fixup_ppaddr before init
add a new init0 section, and initialize it with perl_init.
Needed to set the op_ppaddr when a init function calls an already
compiled function, such as CALLREGCOMP() calling utf8::SWASHINIT
when doing a m//i.
We cannot postpone CALLREGCOMP() to init2, so we need to init the op_ppaddr
before when using -fppaddr/-O1.