C 1.45_14: support utf8 qr
B::PMOP->precomp does not set the utf8 flag, so
we check for utf8 chars in compiled qr strings by ourselves
and set the UTF8 flags then.
Beware that DateTime has % in the $re str, so we cannot sprintf it.

Fixes issue 333 and issue 338, and was checked with qr tests
20 39 44 71 131 143 1431 1432 330 333 338