C: >=5.18 init utf8_tofold swash dynamically to avoid free on static errors
We need this swash with utf8 m//i, but since the &PL_utf8_tofold is empty
we need to init it somehow, to avoid free static pv2... errors (the old swash).
This does not save the folding tables into the binary.
We really need to init utf8_tofold statically (TODO)