C: make -fwalkall the new default. fixes #348
Fixes missing dumped method-only packages.  The order of walking the
packages shalt not influence which packages will not be dumped, when a
dependency gets added later and the package was already marked for deletion.
We don't delete user packages anymore, so we need to walk all of them
exhaustively.

Always save all user-loaded packages. Skip only compiler-loaded packages if
not used by user code.  The compiler loads and dumps now additionally only
the heavy parts for utf8 and bytes, and utf8 for m///i, and AnyDBM for
dbmopen, PerlIO::Scalar for __DATA__ handles, and Coro::State::_jit if
required, but nothing else.