C 1.49_02: skip IO::Handle bloat for *STDOUT
Since 5.18 *STDOUT uses IO::Handle as stash, which is
already properly initialized by PerlIO in core.
Do not need to save it again in *STDOUT->save_magic.
IO::Handle brought in myriads of unneeded dependencies.

This fixes t/e_perlcc tests 21 and 22 which assumed
that an empty script will not add XS modules.