C: Another dirty xs_init hack: detect List::MoreUtils
See #364 why this will not work.
First, we need to detect method_named(PV "bootstrap") and all its args.
The 1st arg will be the XS package to boot, the other args must be added to this method.
Currently we are missing both. It only works because the default case is covered, and
we rarelily use development XS versions (n.nn_nn).