C: 1.52_01 CPANEL_BUILD marks patched Errno and DynaLoader
to reduce Config bloat. See issue #374.
Needs compilation with env CPANEL_BUILD=1 and does not work in
the general case.

For the general case we could think of a new
-fno-config which hardcodes the compile-time Config values when
they are fetched, but this will only run on the same system, with
the very same Config, e.g. paths and settings.