utf8 stash fix for B::GV::save
This commit is fixing one utf8 issue with stash.
View test 372 added to testc.sh to reproduce the error.

Use strlen_flags helper for detecting utf8 strings and
their length. This avoids duplicate logic for unicode strings.

As a bonus this is also fixing previous calls to gv_fetchpvn
with incorrect argument number:
- gv_fetchpvn($name, "GV_ADD", "SVt_PV")
- gv_fetchpvn($name, "GV_ADD", "SVt_PVCV")