Fix travis compilation issues.
Travis needs us to use perl -Mblib when we compile or it can't find B::C. This
is probably an oversight we should have long ago noticed since it means we
weren't always testing what we thought we were.

Based on the use of your prove command, the the c file will include -Mblib prior
to compilation to ensure the desired B::C is used.