Do not use the module Config in B::C
Instead of using Config and bloating compiled code, Simply use the generated
B::C::Flags file to store %Config values. Then do a dirty import in B::C so the
code in B::C does not have to be altered.

This should solve a class of problems related to compiling in Config into
binaries that never use Config.