C: Tie::Hash::NamedCapture linked to unnecessarily
Just the act of looking at the symbol triggers T:H:NC to load,
because of some gv_fetch magic.
So if we simply walk the symbol tables first before hitting the GVs in main,
we can work around this problem.  But overall I think the approach here
needs to be questioned.

Fixes GH #247