Remove memory address check from PVMG::save
Issue #322: Test 3050 is flapping due to an extra optimization in PVMG::save

Removing this optimizations generates no regressions,
and fix this flapping test.

Looks like pretty fragile to check something greater than 0x400000,
we can use some lower values and also need to adjust them.
As we have a check on fullname, this is not so bad.

Simplify the ITHREADS logic, we do not need to check it twice.