C: allow unicode \w symbols in B::GV::savecv
The main stash could contain some unicode functions.
Issue #320: in some cases, unicode functions are not saved
as we are using a regexp to skip them.

Adjust B::GV::savecv to be more permissive.

Rename previous test 320 as 3200,
we are now providing a test case for 320.
Fixes GH #320 and GH #301.

(with fixups -rurban)
Signed-off-by: Reini Urban <rurban@cpanel.net>