B-C: skip perlcc with -m32 cross and PERL_CORE
Mostly does not work, only on darwin with install_name in dylibs.
It works from CPAN setups, because there you can override and fixup
the linker flags. But we don't want to smoke that scenario.

Configure does not probe yet loadable shared libs, only
headers and the default libpaths, but not the libpath needed and
used for the cross compiler.