Fix for "Undefined subroutine utf8::SWASHNEW called at..."
In programs that do not trigger B::C to call load_utf8_heavy() prior to
walksymtable for savecv, you end up with the wrong CV for SWASHNEW. walksymtable
needs to call load_utf8_heavy in those cases. There's probably a more elegant
approach to this but this gets things working until we can do a deeper analysis.

References: CPANEL-6105