cc-harness use correct CORE dirs when in core
In sub cc_harness_msvc, the CC was picking installed C:\perl\lib\CORE
headers, which are a different older perl on my machine, over uninstalled
cperl master headers. This caused CC compile failures.

On Win32 perl, the cperl5**.a|lib is stored in lib/CORE, not the root of
the source tree.